Move type policies
