+++ packages/venia-ui/lib/drivers/adapter.js
@@ -6,8 +6,8 @@ import { ApolloClient } from '@apollo/client/core';
 import { InMemoryCache } from '@apollo/client/cache';
 import { Provider as ReduxProvider } from 'react-redux';
 import { BrowserRouter } from 'react-router-dom';
+import typePolicies from '@magento/peregrine/lib/gql/policies';
 
-import typePolicies from '../policies';
 import { shrinkGETQuery } from '../util/shrinkGETQuery';
 
 /**
